@@ -1,236 +1,218 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="Program.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   A program to setup the database.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.DatabaseSetup
+namespace NetCoreMQTTExampleCluster.DatabaseSetup;
+
+/// <summary>
+/// A program to setup the database.
+/// </summary>
+public static class Program
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
-    using System.IO;
-    using System.Linq;
-    using System.Reflection;
-    using System.Threading.Tasks;
-
-    using Microsoft.AspNetCore.Identity;
-
-    using NetCoreMQTTExampleCluster.Storage;
-    using NetCoreMQTTExampleCluster.Storage.Data;
-    using NetCoreMQTTExampleCluster.Storage.Repositories.Implementation;
-    using NetCoreMQTTExampleCluster.Storage.Repositories.Interfaces;
-
-    using Newtonsoft.Json;
+    /// <summary>
+    /// The database version repository.
+    /// </summary>
+    private static IDatabaseVersionRepository databaseVersionRepository;
 
     /// <summary>
-    ///     A program to setup the database.
+    /// The whitelist repository.
     /// </summary>
-    public static class Program
+    private static WhitelistRepository whitelistRepository;
+
+    /// <summary>
+    /// The user repository.
+    /// </summary>
+    private static IUserRepository userRepository;
+
+    /// <summary>
+    /// The database helper.
+    /// </summary>
+    private static IDatabaseHelper databaseHelper;
+
+    /// <summary>
+    /// The main method of the program.
+    /// </summary>
+    /// <returns>A <see cref="Task" />.</returns>
+    public static async Task Main()
     {
-        /// <summary>
-        ///     The database version repository.
-        /// </summary>
-        private static IDatabaseVersionRepository databaseVersionRepository;
+        var connectionSettings = await ReadSettingsFile();
 
-        /// <summary>
-        /// The whitelist repository.
-        /// </summary>
-        private static WhitelistRepository whitelistRepository;
+        userRepository = new UserRepository(connectionSettings);
+        databaseVersionRepository = new DatabaseVersionRepository(connectionSettings);
+        whitelistRepository = new WhitelistRepository(connectionSettings);
+        databaseHelper = new DatabaseHelper(connectionSettings);
 
-        /// <summary>
-        ///     The user repository.
-        /// </summary>
-        private static IUserRepository userRepository;
+        Console.WriteLine("Setting up the database...");
 
-        /// <summary>
-        ///     The database helper.
-        /// </summary>
-        private static IDatabaseHelper databaseHelper;
+        Console.WriteLine("Deleting database...");
+        await databaseHelper.DeleteDatabase(connectionSettings.Database);
 
-        /// <summary>
-        ///     The main method of the program.
-        /// </summary>
-        /// <returns>A <see cref="Task" />.</returns>
-        public static async Task Main()
+        Console.WriteLine("Creating database...");
+        await databaseHelper.CreateDatabase(connectionSettings.Database);
+
+        Console.WriteLine("Adding TimeScaleDB extension...");
+        await databaseHelper.EnableTimeScaleDbExtension();
+
+        Console.WriteLine("Adding Orleans tables...");
+        await databaseHelper.CreateOrleansTables();
+
+        Console.WriteLine("Creating tables...");
+        await databaseHelper.CreateAllTables(true);
+
+        Console.WriteLine("Creating hyper tables...");
+        await databaseHelper.CreateHyperTables();
+
+        Console.WriteLine("Creating compound index...");
+        await databaseHelper.CreateCompoundIndex();
+
+        Console.WriteLine("Adding seed data...");
+        await SeedData();
+
+        Console.WriteLine("Press any key to close this window...");
+        Console.ReadKey();
+    }
+
+    /// <summary>
+    /// Reads the settings file.
+    /// </summary>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task<MqttDatabaseConnectionSettings> ReadSettingsFile()
+    {
+        var currentLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+        var settingsFile = Path.Combine(currentLocation, "appsettings.json");
+        var settingsString = await File.ReadAllTextAsync(settingsFile);
+        return JsonConvert.DeserializeObject<MqttDatabaseConnectionSettings>(settingsString);
+    }
+
+    /// <summary>
+    /// Seeds some data to the database.
+    /// </summary>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task SeedData()
+    {
+        await InsertDatabaseVersion();
+        var (user1Id, user2Id) = await InsertUsers();
+        await AddWhiteAndBlackListsForFirstUser(user1Id);
+        await AddWhiteAndBlackListsForSecondUser(user2Id);
+    }
+
+    /// <summary>
+    /// Adds the blacklist and whitelist items for the first user.
+    /// </summary>
+    /// <param name="userId">The user identifier.</param>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task AddWhiteAndBlackListsForFirstUser(Guid userId)
+    {
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            var connectionSettings = await ReadSettingsFile();
+            Type = BlacklistWhitelistType.Publish,
+            UserId = userId,
+            Value = "a/b"
+        });
 
-            userRepository = new UserRepository(connectionSettings);
-            databaseVersionRepository = new DatabaseVersionRepository(connectionSettings);
-            whitelistRepository = new WhitelistRepository(connectionSettings);
-            databaseHelper = new DatabaseHelper(connectionSettings);
-
-            Console.WriteLine("Setting up the database...");
-
-            Console.WriteLine("Deleting database...");
-            await databaseHelper.DeleteDatabase(connectionSettings.Database);
-
-            Console.WriteLine("Creating database...");
-            await databaseHelper.CreateDatabase(connectionSettings.Database);
-
-            Console.WriteLine("Adding TimeScaleDB extension...");
-            await databaseHelper.EnableTimeScaleDbExtension();
-
-            Console.WriteLine("Adding Orleans tables...");
-            await databaseHelper.CreateOrleansTables();
-
-            Console.WriteLine("Creating tables...");
-            await databaseHelper.CreateAllTables(true);
-
-            Console.WriteLine("Creating hyper tables...");
-            await databaseHelper.CreateHyperTables();
-
-            Console.WriteLine("Creating compound index...");
-            await databaseHelper.CreateCompoundIndex();
-
-            Console.WriteLine("Adding seed data...");
-            await SeedData();
-
-            Console.WriteLine("Press any key to close this window...");
-            Console.ReadKey();
-        }
-
-        /// <summary>
-        /// Reads the settings file.
-        /// </summary>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task<MqttDatabaseConnectionSettings> ReadSettingsFile()
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            var currentLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            var settingsFile = Path.Combine(currentLocation, "appsettings.json");
-            var settingsString = await File.ReadAllTextAsync(settingsFile);
-            return JsonConvert.DeserializeObject<MqttDatabaseConnectionSettings>(settingsString);
-        }
+            Type = BlacklistWhitelistType.Publish,
+            UserId = userId,
+            Value = "a/d"
+        });
 
-        /// <summary>
-        /// Seeds some data to the database.
-        /// </summary>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task SeedData()
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            await InsertDatabaseVersion();
-            var (user1Id, user2Id) = await InsertUsers();
-            await AddWhiteAndBlackListsForFirstUser(user1Id);
-            await AddWhiteAndBlackListsForSecondUser(user2Id);
-        }
+            Type = BlacklistWhitelistType.Subscribe,
+            UserId = userId,
+            Value = "d/e"
+        });
 
-        /// <summary>
-        /// Adds the blacklist and whitelist items for the first user.
-        /// </summary>
-        /// <param name="userId">The user identifier.</param>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task AddWhiteAndBlackListsForFirstUser(Guid userId)
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Publish,
-                UserId = userId,
-                Value = "a/b"
-            });
+            Type = BlacklistWhitelistType.Subscribe,
+            UserId = userId,
+            Value = "d"
+        });
+    }
 
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Publish,
-                UserId = userId,
-                Value = "a/d"
-            });
-
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Subscribe,
-                UserId = userId,
-                Value = "d/e"
-            });
-
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Subscribe,
-                UserId = userId,
-                Value = "d"
-            });
-        }
-
-        /// <summary>
-        /// Adds the blacklist and whitelist items for the second user: mqtt-broker-sync.
-        /// </summary>
-        /// <param name="userId">The user identifier.</param>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task AddWhiteAndBlackListsForSecondUser(Guid userId)
+    /// <summary>
+    /// Adds the blacklist and whitelist items for the second user: mqtt-broker-sync.
+    /// </summary>
+    /// <param name="userId">The user identifier.</param>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task AddWhiteAndBlackListsForSecondUser(Guid userId)
+    {
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Publish,
-                UserId = userId,
-                Value = "#"
-            });
+            Type = BlacklistWhitelistType.Publish,
+            UserId = userId,
+            Value = "#"
+        });
 
-            await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
-            {
-                Type = BlacklistWhitelistType.Subscribe,
-                UserId = userId,
-                Value = "#"
-            });
-        }
-
-        /// <summary>
-        /// Inserts the <see cref="User"/>s.
-        /// </summary>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task<(Guid, Guid)> InsertUsers()
+        await whitelistRepository.InsertWhitelistItem(new BlacklistWhitelist
         {
-            var passwordHasher = new PasswordHasher<User>();
+            Type = BlacklistWhitelistType.Subscribe,
+            UserId = userId,
+            Value = "#"
+        });
+    }
 
-            // Add test user
-            var userId = Guid.NewGuid();
+    /// <summary>
+    /// Inserts the <see cref="User"/>s.
+    /// </summary>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task<(Guid, Guid)> InsertUsers()
+    {
+        var passwordHasher = new PasswordHasher<User>();
 
-            var user = new User
-            {
-                Id = userId,
-                UserName = "Test",
-                ValidateClientId = false
-            };
+        // Add test user
+        var userId = Guid.NewGuid();
 
-            user.PasswordHash = passwordHasher.HashPassword(user, "test");
-            await userRepository.InsertUser(user);
-
-            // Add broker user
-            var user2Id = Guid.NewGuid();
-
-            var user2 = new User
-            {
-                Id = user2Id,
-                UserName = "mqtt-broker-sync",
-                ClientIdPrefix = "mqtt-broker-sync",
-                ValidateClientId = true,
-                IsSyncUser = true
-            };
-
-            user2.PasswordHash = passwordHasher.HashPassword(user2, "Test");
-            await userRepository.InsertUser(user2);
-            return (userId, user2Id);
-        }
-
-        /// <summary>
-        /// Inserts the <see cref="DatabaseVersion"/>.
-        /// </summary>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        private static async Task InsertDatabaseVersion()
+        var user = new User
         {
-            var databaseVersions = await databaseVersionRepository.GetDatabaseVersions() ?? new List<DatabaseVersion>();
+            Id = userId,
+            UserName = "Test",
+            ValidateClientId = false
+        };
 
-            if (!databaseVersions.Any())
-            {
-                await databaseVersionRepository.InsertDatabaseVersion(
-                    new DatabaseVersion
-                    {
-                        Name = "1.0.0",
-                        Number = 100
-                    });
-            }
+        user.PasswordHash = passwordHasher.HashPassword(user, "test");
+        await userRepository.InsertUser(user);
+
+        // Add broker user
+        var user2Id = Guid.NewGuid();
+
+        var user2 = new User
+        {
+            Id = user2Id,
+            UserName = "mqtt-broker-sync",
+            ClientIdPrefix = "mqtt-broker-sync",
+            ValidateClientId = true,
+            IsSyncUser = true
+        };
+
+        user2.PasswordHash = passwordHasher.HashPassword(user2, "Test");
+        await userRepository.InsertUser(user2);
+        return (userId, user2Id);
+    }
+
+    /// <summary>
+    /// Inserts the <see cref="DatabaseVersion"/>.
+    /// </summary>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    private static async Task InsertDatabaseVersion()
+    {
+        var databaseVersions = await databaseVersionRepository.GetDatabaseVersions() ?? new List<DatabaseVersion>();
+
+        if (!databaseVersions.Any())
+        {
+            await databaseVersionRepository.InsertDatabaseVersion(
+                new DatabaseVersion
+                {
+                    Name = "1.0.0",
+                    Number = 100
+                });
         }
     }
 }
