@@ -1,222 +1,203 @@
 // --------------------------------------------------------------------------------------------------------------------
 // <copyright file="MqttClientGrain.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   The grain for one client identifier.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.Grains
+namespace NetCoreMQTTExampleCluster.Grains;
+
+/// <inheritdoc cref="IMqttClientGrain" />
+/// <summary>
+/// The grain for one client identifier.
+/// </summary>
+/// <seealso cref="IMqttClientGrain" />
+public class MqttClientGrain : Grain, IMqttClientGrain
 {
-    using System;
-    using System.Diagnostics.CodeAnalysis;
-    using System.Threading.Tasks;
+    /// <summary>
+    /// Gets or sets the data limit cache for throttling for monthly data.
+    /// </summary>
+    private static readonly MemoryCache DataLimitCacheMonth = new(new MemoryCacheOptions());
 
-    using Microsoft.AspNetCore.Identity;
-    using Microsoft.Extensions.Caching.Memory;
+    /// <summary>
+    /// The <see cref="IPasswordHasher{TUser}" />.
+    /// </summary>
+    private static readonly IPasswordHasher<User> PasswordHasher = new PasswordHasher<User>();
 
-    using MQTTnet.Server;
+    /// <summary>
+    /// The user repository.
+    /// </summary>
+    private readonly IUserRepository userRepository;
 
-    using NetCoreMQTTExampleCluster.Grains.Interfaces;
-    using NetCoreMQTTExampleCluster.Storage.Data;
-    using NetCoreMQTTExampleCluster.Storage.Repositories.Interfaces;
-    using NetCoreMQTTExampleCluster.Validation;
+    /// <summary>
+    /// The MQTT validator.
+    /// </summary>
+    private readonly IMqttValidator mqttValidator;
 
-    using Orleans;
+    /// <summary>
+    /// The user.
+    /// </summary>
+    private User user;
 
-    using Serilog;
+    /// <summary>
+    /// The user data.
+    /// </summary>
+    private UserData userData = new();
+
+    /// <summary>
+    /// A value indicating whether the caches are loaded or not.
+    /// </summary>
+    private bool cacheLoaded;
+
+    /// <summary>
+    /// The client identifier.
+    /// </summary>
+    private string clientId = string.Empty;
+
+    /// <summary>
+    /// The logger.
+    /// </summary>
+    private ILogger logger;
 
     /// <inheritdoc cref="IMqttClientGrain" />
     /// <summary>
-    ///     The grain for one client identifier.
+    /// Initializes a new instance of the <see cref="MqttClientGrain" /> class.
     /// </summary>
+    /// <param name="userRepository">The user repository.</param>
+    /// <param name="mqttValidator">The MQTT validator.</param>
     /// <seealso cref="IMqttClientGrain" />
-    public class MqttClientGrain : Grain, IMqttClientGrain
+    public MqttClientGrain(IUserRepository userRepository, IMqttValidator mqttValidator)
     {
-        /// <summary>
-        ///     Gets or sets the data limit cache for throttling for monthly data.
-        /// </summary>
-        private static readonly MemoryCache DataLimitCacheMonth = new MemoryCache(new MemoryCacheOptions());
+        this.userRepository = userRepository;
+        this.mqttValidator = mqttValidator;
+        this.logger = Log.ForContext("Grain", nameof(MqttClientGrain));
+    }
 
-        /// <summary>
-        ///     The <see cref="IPasswordHasher{TUser}" />.
-        /// </summary>
-        private static readonly IPasswordHasher<User> PasswordHasher = new PasswordHasher<User>();
+    /// <inheritdoc cref="Grain" />
+    /// <summary>
+    /// This method is called at the end of the process of activating a grain.
+    /// It is called before any messages have been dispatched to the grain.
+    /// For grains with declared persistent state, this method is called after the State property has been populated.
+    /// </summary>
+    /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
+    /// <seealso cref="Grain" />
+    public override Task OnActivateAsync()
+    {
+        this.clientId = this.GetPrimaryKeyString();
+        this.logger = Log.ForContext("Grain", nameof(MqttClientGrain)).ForContext("Id", this.clientId);
+        return base.OnActivateAsync();
+    }
 
-        /// <summary>
-        ///     The user repository.
-        /// </summary>
-        private readonly IUserRepository userRepository;
-
-        /// <summary>
-        /// The MQTT validator.
-        /// </summary>
-        private readonly IMqttValidator mqttValidator;
-
-        /// <summary>
-        /// The user.
-        /// </summary>
-        private User user;
-
-        /// <summary>
-        /// The user data.
-        /// </summary>
-        private UserData userData = new UserData();
-
-        /// <summary>
-        /// A value indicating whether the caches are loaded or not.
-        /// </summary>
-        private bool cacheLoaded;
-
-        /// <summary>
-        ///     The client identifier.
-        /// </summary>
-        private string clientId = string.Empty;
-
-        /// <summary>
-        ///     The logger.
-        /// </summary>
-        private ILogger logger;
-
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="MqttClientGrain" /> class.
-        /// </summary>
-        /// <param name="userRepository">The user repository.</param>
-        /// <param name="mqttValidator">The MQTT validator.</param>
-        /// <seealso cref="IMqttClientGrain" />
-        public MqttClientGrain(IUserRepository userRepository, IMqttValidator mqttValidator)
+    /// <inheritdoc cref="IMqttClientGrain" />
+    /// <summary>
+    /// Proceeds the connection for one client identifier.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <returns>A value indicating whether the connection is accepted or not.</returns>
+    /// <seealso cref="IMqttClientGrain" />
+    public async Task<bool> ProceedConnect(SimpleMqttConnectionValidatorContext context)
+    {
+        try
         {
-            this.userRepository = userRepository;
-            this.mqttValidator = mqttValidator;
-            this.logger = Log.ForContext("Grain", nameof(MqttClientGrain));
-        }
+            this.user = await this.userRepository.GetUserByName(context.UserName);
 
-        /// <inheritdoc cref="Grain" />
-        /// <summary>
-        ///     This method is called at the end of the process of activating a grain.
-        ///     It is called before any messages have been dispatched to the grain.
-        ///     For grains with declared persistent state, this method is called after the State property has been populated.
-        /// </summary>
-        /// <returns>A <see cref="Task" /> representing any asynchronous operation.</returns>
-        /// <seealso cref="Grain" />
-        public override Task OnActivateAsync()
-        {
-            this.clientId = this.GetPrimaryKeyString();
-            this.logger = Log.ForContext("Grain", nameof(MqttClientGrain)).ForContext("Id", this.clientId);
-            return base.OnActivateAsync();
-        }
-
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        ///     Proceeds the connection for one client identifier.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <returns>A value indicating whether the connection is accepted or not.</returns>
-        /// <seealso cref="IMqttClientGrain" />
-        public async Task<bool> ProceedConnect(SimpleMqttConnectionValidatorContext context)
-        {
-            try
+            if (this.user is null)
             {
-                this.user = await this.userRepository.GetUserByName(context.UserName);
-
-                if (this.user == null)
-                {
-                    return false;
-                }
-
-                await this.RefreshCache(false);
-                return this.mqttValidator.ValidateConnection(context, this.user, PasswordHasher);
-            }
-            catch (Exception ex)
-            {
-                this.logger.Error("An error occurred: {@ex}.", ex);
                 return false;
             }
-        }
 
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        ///     Proceeds the published message for one client identifier.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <returns>A value indicating whether the published message is accepted or not.</returns>
-        /// <seealso cref="IMqttClientGrain" />
-        public Task<bool> ProceedPublish(MqttApplicationMessageInterceptorContext context)
+            await this.RefreshCache(false);
+            return this.mqttValidator.ValidateConnection(context, this.user, PasswordHasher);
+        }
+        catch (Exception ex)
         {
-            try
-            {
-                var result = this.mqttValidator.ValidatePublish(context, this.userData.PublishBlacklist, this.userData.PublishWhitelist, this.user, DataLimitCacheMonth, this.userData.ClientIdPrefixes);
-                return Task.FromResult(result);
-            }
-            catch (Exception ex)
-            {
-                this.logger.Error("An error occurred: {@ex}.", ex);
-                return Task.FromResult(false);
-            }
+            this.logger.Error("An error occurred: {@ex}.", ex);
+            return false;
         }
+    }
 
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        ///     Proceeds the subscription for one client identifier.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <returns>A value indicating whether the subscription is accepted or not.</returns>
-        /// <seealso cref="IMqttClientGrain" />
-        public Task<bool> ProceedSubscription(MqttSubscriptionInterceptorContext context)
+    /// <inheritdoc cref="IMqttClientGrain" />
+    /// <summary>
+    /// Proceeds the published message for one client identifier.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <returns>A value indicating whether the published message is accepted or not.</returns>
+    /// <seealso cref="IMqttClientGrain" />
+    public Task<bool> ProceedPublish(MqttApplicationMessageInterceptorContext context)
+    {
+        try
         {
-            try
-            {
-                var result = this.mqttValidator.ValidateSubscription(context, this.userData.SubscriptionBlacklist, this.userData.SubscriptionWhitelist, this.user, this.userData.ClientIdPrefixes);
-                return Task.FromResult(result);
-            }
-            catch (Exception ex)
-            {
-                this.logger.Error("An error occurred: {@ex}.", ex);
-                return Task.FromResult(false);
-            }
+            var result = this.mqttValidator.ValidatePublish(context, this.userData.PublishBlacklist, this.userData.PublishWhitelist, this.user, DataLimitCacheMonth, this.userData.ClientIdPrefixes);
+            return Task.FromResult(result);
         }
-
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        /// Checks whether the user is a user used for synchronization.
-        /// </summary>
-        /// <returns>A value indicating whether the user is a broker user or not.</returns>
-        /// <seealso cref="IMqttClientGrain" />
-        public Task<bool> IsUserBrokerUser()
+        catch (Exception ex)
         {
-            try
-            {
-                var result = this.user.IsSyncUser;
-                return Task.FromResult(result);
-            }
-            catch (Exception ex)
-            {
-                this.logger.Error("An error occurred: {@ex}.", ex);
-                return Task.FromResult(false);
-            }
+            this.logger.Error("An error occurred: {@ex}.", ex);
+            return Task.FromResult(false);
         }
+    }
 
-        /// <inheritdoc cref="IMqttClientGrain" />
-        /// <summary>
-        /// Tells the grain to refresh its cache.
-        /// </summary>
-        /// <param name="force">Forces a cache update.</param>
-        /// <returns>A <see cref="Task"/> representing any asynchronous operation.</returns>
-        /// <seealso cref="IMqttClientGrain" />
-        public async Task RefreshCache(bool force)
+    /// <inheritdoc cref="IMqttClientGrain" />
+    /// <summary>
+    /// Proceeds the subscription for one client identifier.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <returns>A value indicating whether the subscription is accepted or not.</returns>
+    /// <seealso cref="IMqttClientGrain" />
+    public Task<bool> ProceedSubscription(MqttSubscriptionInterceptorContext context)
+    {
+        try
         {
-            if (!force)
-            {
-                if (this.cacheLoaded)
-                {
-                    return;
-                }
-            }
-
-            this.userData = await this.userRepository.GetUserData(this.user.Id);
-            this.cacheLoaded = true;
+            var result = this.mqttValidator.ValidateSubscription(context, this.userData.SubscriptionBlacklist, this.userData.SubscriptionWhitelist, this.user, this.userData.ClientIdPrefixes);
+            return Task.FromResult(result);
         }
+        catch (Exception ex)
+        {
+            this.logger.Error("An error occurred: {@ex}.", ex);
+            return Task.FromResult(false);
+        }
+    }
+
+    /// <inheritdoc cref="IMqttClientGrain" />
+    /// <summary>
+    /// Checks whether the user is a user used for synchronization.
+    /// </summary>
+    /// <returns>A value indicating whether the user is a broker user or not.</returns>
+    /// <seealso cref="IMqttClientGrain" />
+    public Task<bool> IsUserBrokerUser()
+    {
+        try
+        {
+            var result = this.user.IsSyncUser;
+            return Task.FromResult(result);
+        }
+        catch (Exception ex)
+        {
+            this.logger.Error("An error occurred: {@ex}.", ex);
+            return Task.FromResult(false);
+        }
+    }
+
+    /// <inheritdoc cref="IMqttClientGrain" />
+    /// <summary>
+    /// Tells the grain to refresh its cache.
+    /// </summary>
+    /// <param name="force">Forces a cache update.</param>
+    /// <returns>A <see cref="Task"/> representing any asynchronous operation.</returns>
+    /// <seealso cref="IMqttClientGrain" />
+    public async Task RefreshCache(bool force)
+    {
+        if (!force)
+        {
+            if (this.cacheLoaded)
+            {
+                return;
+            }
+        }
+
+        this.userData = await this.userRepository.GetUserData(this.user.Id);
+        this.cacheLoaded = true;
     }
 }
