@@ -1,31 +1,27 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="BlacklistWhitelistType.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   An enumeration representing the available blacklist or whitelist types.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.Storage.Data
+namespace NetCoreMQTTExampleCluster.Storage.Data;
+
+/// <summary>
+/// An enumeration representing the available blacklist or whitelist types.
+/// </summary>
+[JsonConverter(typeof(StringEnumConverter))]
+public enum BlacklistWhitelistType
 {
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Converters;
+    /// <summary>
+    /// The subscribe blacklist or whitelist type.
+    /// </summary>
+    Subscribe,
 
     /// <summary>
-    ///     An enumeration representing the available blacklist or whitelist types.
+    /// The publish blacklist or whitelist type.
     /// </summary>
-    [JsonConverter(typeof(StringEnumConverter))]
-    public enum BlacklistWhitelistType
-    {
-        /// <summary>
-        ///     The subscribe blacklist or whitelist type.
-        /// </summary>
-        Subscribe,
-
-        /// <summary>
-        ///     The publish blacklist or whitelist type.
-        /// </summary>
-        Publish
-    }
+    Publish
 }
