@@ -1,50 +1,47 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="PublishedMessagePayload.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   A class to serialize n published message payload to the JSON binary field in the database.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.Storage.Data
+namespace NetCoreMQTTExampleCluster.Storage.Data;
+
+/// <summary>
+/// A class to serialize n published message payload to the JSON binary field in the database.
+/// </summary>
+public class PublishedMessagePayload
 {
-    using Newtonsoft.Json;
+    /// <summary>
+    /// Initializes a new instance of the <see cref="PublishedMessagePayload"/> class.
+    /// This constructor is needed for JSON (de-)serialization.
+    /// </summary>
+    public PublishedMessagePayload()
+    {
+    }
 
     /// <summary>
-    /// A class to serialize n published message payload to the JSON binary field in the database.
+    /// Initializes a new instance of the <see cref="PublishedMessagePayload"/> class.
     /// </summary>
-    public class PublishedMessagePayload
+    /// <param name="message">The message.</param>
+    public PublishedMessagePayload(string message)
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="PublishedMessagePayload"/> class.
-        /// This constructor is needed for JSON (de-)serialization.
-        /// </summary>
-        public PublishedMessagePayload()
-        {
-        }
+        this.Message = message;
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="PublishedMessagePayload"/> class.
-        /// </summary>
-        /// <param name="message">The message.</param>
-        public PublishedMessagePayload(string message)
-        {
-            this.Message = message;
-        }
+    /// <summary>
+    /// Gets or sets the message payload.
+    /// </summary>
+    public string Message { get; set; } = string.Empty;
 
-        /// <summary>
-        /// Gets or sets the message payload.
-        /// </summary>
-        public string Message { get; set; }
-
-        /// <summary>
-        ///     Returns a <see cref="string"></see> representation of the <see cref="PublishedMessagePayload" /> class.
-        /// </summary>
-        /// <returns>A <see cref="string"></see> representation of the <see cref="PublishedMessagePayload" /> class.</returns>
-        public override string ToString()
-        {
-            return JsonConvert.SerializeObject(this);
-        }
+    /// <summary>
+    /// Returns a <see cref="string"></see> representation of the <see cref="PublishedMessagePayload" /> class.
+    /// </summary>
+    /// <returns>A <see cref="string"></see> representation of the <see cref="PublishedMessagePayload" /> class.</returns>
+    public override string ToString()
+    {
+        return JsonConvert.SerializeObject(this);
     }
 }
