@@ -1,60 +1,57 @@
 // --------------------------------------------------------------------------------------------------------------------
 // <copyright file="TopicCheckerTestsCrossOperator.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   A test class to test the <see cref="TopicChecker" /> with the # operator.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.TopicCheck.Tests
+namespace NetCoreMQTTExampleCluster.TopicCheck.Tests;
+
+/// <summary>
+/// A test class to test the <see cref="TopicChecker" /> with the # operator.
+/// </summary>
+[TestClass]
+public class TopicCheckerTestsCrossOperator
 {
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
+    /// <summary>
+    /// Checks the tester with a valid topic for the # operator.
+    /// </summary>
+    [TestMethod]
+    public void CheckSingleValueCrossMatch()
+    {
+        var result = TopicChecker.Regex("a/#", "a/b");
+        Assert.IsTrue(result);
+    }
 
     /// <summary>
-    ///     A test class to test the <see cref="TopicChecker" /> with the # operator.
+    /// Checks the tester with another valid topic for the # operator.
     /// </summary>
-    [TestClass]
-    public class TopicCheckerTestsCrossOperator
+    [TestMethod]
+    public void CheckSingleValueCrossMatch2()
     {
-        /// <summary>
-        ///     Checks the tester with a valid topic for the # operator.
-        /// </summary>
-        [TestMethod]
-        public void CheckSingleValueCrossMatch()
-        {
-            var result = TopicChecker.Regex("a/#", "a/b");
-            Assert.IsTrue(result);
-        }
+        var result = TopicChecker.Regex("a/#", "a/b/c");
+        Assert.IsTrue(result);
+    }
 
-        /// <summary>
-        ///     Checks the tester with another valid topic for the # operator.
-        /// </summary>
-        [TestMethod]
-        public void CheckSingleValueCrossMatch2()
-        {
-            var result = TopicChecker.Regex("a/#", "a/b/c");
-            Assert.IsTrue(result);
-        }
+    /// <summary>
+    /// Checks the tester with a valid topic with a # for the # operator.
+    /// </summary>
+    [TestMethod]
+    public void CheckSingleValueCrossMatchWithCross()
+    {
+        var result = TopicChecker.Regex("a/#", "a/#");
+        Assert.IsTrue(result);
+    }
 
-        /// <summary>
-        ///     Checks the tester with a valid topic with a # for the # operator.
-        /// </summary>
-        [TestMethod]
-        public void CheckSingleValueCrossMatchWithCross()
-        {
-            var result = TopicChecker.Regex("a/#", "a/#");
-            Assert.IsTrue(result);
-        }
-
-        /// <summary>
-        ///     Checks the tester with a valid topic with a + for the # operator.
-        /// </summary>
-        [TestMethod]
-        public void CheckSingleValueCrossMatchWithPlus()
-        {
-            var result = TopicChecker.Regex("a/#", "a/+");
-            Assert.IsTrue(result);
-        }
+    /// <summary>
+    /// Checks the tester with a valid topic with a + for the # operator.
+    /// </summary>
+    [TestMethod]
+    public void CheckSingleValueCrossMatchWithPlus()
+    {
+        var result = TopicChecker.Regex("a/#", "a/+");
+        Assert.IsTrue(result);
     }
 }
