@@ -1,347 +1,330 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
+// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="MqttValidator.cs" company="Hämmer Electronics">
-//   Copyright (c) 2020 All rights reserved.
+//   Copyright (c) All rights reserved.
 // </copyright>
 // <summary>
 //   A class to validate the different MQTT contexts.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace NetCoreMQTTExampleCluster.Validation
+namespace NetCoreMQTTExampleCluster.Validation;
+
+/// <inheritdoc cref="IMqttValidator"/>
+/// <summary>
+/// A class to validate the different MQTT contexts.
+/// </summary>
+/// <seealso cref="IMqttValidator"/>
+public class MqttValidator : IMqttValidator
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-
-    using Microsoft.AspNetCore.Identity;
-    using Microsoft.Extensions.Caching.Memory;
-
-    using MQTTnet.Server;
-
-    using NetCoreMQTTExampleCluster.Grains.Interfaces;
-    using NetCoreMQTTExampleCluster.Models.Extensions;
-    using NetCoreMQTTExampleCluster.Storage.Data;
-    using NetCoreMQTTExampleCluster.TopicCheck;
-
-    using Serilog;
+    /// <summary>
+    /// The logger.
+    /// </summary>
+    private static readonly ILogger Logger = Log.ForContext<MqttValidator>();
 
     /// <inheritdoc cref="IMqttValidator"/>
     /// <summary>
-    /// A class to validate the different MQTT contexts.
+    /// Validates the connection.
     /// </summary>
+    /// <param name="context">The context.</param>
+    /// <param name="user">The user.</param>
+    /// <param name="passwordHasher">The password hasher.</param>
+    /// <returns>A value indicating whether the connection is accepted or not.</returns>
     /// <seealso cref="IMqttValidator"/>
-    public class MqttValidator : IMqttValidator
+    public bool ValidateConnection(SimpleMqttConnectionValidatorContext context, User user, IPasswordHasher<User> passwordHasher)
     {
-        /// <summary>
-        /// The logger.
-        /// </summary>
-        private static readonly ILogger Logger = Log.ForContext<MqttValidator>();
+        Logger.Debug("Executed ValidateConnection with parameters: {@context}, {@user}.", context, user);
 
-        /// <inheritdoc cref="IMqttValidator"/>
-        /// <summary>
-        ///     Validates the connection.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <param name="user">The user.</param>
-        /// <param name="passwordHasher">The password hasher.</param>
-        /// <returns>A value indicating whether the connection is accepted or not.</returns>
-        /// <seealso cref="IMqttValidator"/>
-        public bool ValidateConnection(SimpleMqttConnectionValidatorContext context, User user, IPasswordHasher<User> passwordHasher)
+        Logger.Debug("Current user is {@currentUser}.", user);
+
+        if (user is null)
         {
-            Logger.Debug("Executed ValidateConnection with parameters: {@context}, {@user}.", context, user);
+            Logger.Debug("Current user was null.");
+            return false;
+        }
 
-            Logger.Debug("Current user is {@currentUser}.", user);
+        if (context.UserName != user.UserName)
+        {
+            Logger.Debug("User name in context doesn't match the current user.");
+            return false;
+        }
 
-            if (user == null)
-            {
-                Logger.Debug("Current user was null.");
-                return false;
-            }
+        var hashingResult = passwordHasher.VerifyHashedPassword(user, user.PasswordHash, context.Password);
 
-            if (context.UserName != user.UserName)
-            {
-                Logger.Debug("User name in context doesn't match the current user.");
-                return false;
-            }
+        if (hashingResult == PasswordVerificationResult.Failed)
+        {
+            Logger.Debug("Password verification failed.");
+            return false;
+        }
 
-            var hashingResult = passwordHasher.VerifyHashedPassword(user, user.PasswordHash, context.Password);
-
-            if (hashingResult == PasswordVerificationResult.Failed)
-            {
-                Logger.Debug("Password verification failed.");
-                return false;
-            }
-
-            if (!user.ValidateClientId)
-            {
-                Logger.Debug("Connection valid for {@clientId} and {@user}.", context.ClientId, user);
-                return true;
-            }
-
-            if (string.IsNullOrWhiteSpace(user.ClientIdPrefix))
-            {
-                if (context.ClientId != user.ClientId)
-                {
-                    Logger.Debug("Client id in context doesn't match the current user's client id.");
-                    return false;
-                }
-
-                Logger.Debug("Connection valid for {@clientId} and {@user} when client id prefix was null.", context.ClientId, user);
-            }
-            else
-            {
-                Logger.Debug("Connection valid for {@clientIdPrefix} and {@user} when client id prefix was not null.", user.ClientIdPrefix, user);
-            }
-
+        if (!user.ValidateClientId)
+        {
+            Logger.Debug("Connection valid for {@clientId} and {@user}.", context.ClientId, user);
             return true;
         }
 
-        /// <inheritdoc cref="IMqttValidator"/>
-        /// <summary>
-        ///     Validates the message publication.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <param name="blacklist">The blacklist.</param>
-        /// <param name="whitelist">The whitelist.</param>
-        /// <param name="user">The user.</param>
-        /// <param name="dataLimitCacheMonth">The data limit cache for the month.</param>
-        /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
-        /// <returns>A value indicating whether the published message is accepted or not.</returns>
-        /// <seealso cref="IMqttValidator"/>
-        public bool ValidatePublish(
-            MqttApplicationMessageInterceptorContext context,
-            List<BlacklistWhitelist> blacklist,
-            List<BlacklistWhitelist> whitelist,
-            User user,
-            IMemoryCache dataLimitCacheMonth,
-            List<string> clientIdPrefixes)
+        if (string.IsNullOrWhiteSpace(user.ClientIdPrefix))
         {
-            Logger.Debug("Executed ValidatePublish with parameters: {@context}, {@user}.", context, user);
-
-            var clientIdPrefix = GetClientIdPrefix(context.ClientId, clientIdPrefixes);
-
-            Logger.Debug("Client id prefix is {@clientIdPrefix}.", clientIdPrefix);
-
-            if (user == null)
+            if (context.ClientId != user.ClientId)
             {
-                Logger.Debug("Current user was null.");
+                Logger.Debug("Client id in context doesn't match the current user's client id.");
                 return false;
             }
 
-            var topic = context.ApplicationMessage.Topic;
+            Logger.Debug("Connection valid for {@clientId} and {@user} when client id prefix was null.", context.ClientId, user);
+        }
+        else
+        {
+            Logger.Debug("Connection valid for {@clientIdPrefix} and {@user} when client id prefix was not null.", user.ClientIdPrefix, user);
+        }
 
-            Logger.Debug("Topic was {@topic}.", topic);
+        return true;
+    }
 
-            if (user.ThrottleUser)
+    /// <inheritdoc cref="IMqttValidator"/>
+    /// <summary>
+    /// Validates the message publication.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <param name="blacklist">The blacklist.</param>
+    /// <param name="whitelist">The whitelist.</param>
+    /// <param name="user">The user.</param>
+    /// <param name="dataLimitCacheMonth">The data limit cache for the month.</param>
+    /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
+    /// <returns>A value indicating whether the published message is accepted or not.</returns>
+    /// <seealso cref="IMqttValidator"/>
+    public bool ValidatePublish(
+        MqttApplicationMessageInterceptorContext context,
+        List<BlacklistWhitelist> blacklist,
+        List<BlacklistWhitelist> whitelist,
+        User user,
+        IMemoryCache dataLimitCacheMonth,
+        List<string> clientIdPrefixes)
+    {
+        Logger.Debug("Executed ValidatePublish with parameters: {@context}, {@user}.", context, user);
+
+        var clientIdPrefix = GetClientIdPrefix(context.ClientId, clientIdPrefixes);
+
+        Logger.Debug("Client id prefix is {@clientIdPrefix}.", clientIdPrefix);
+
+        if (user is null)
+        {
+            Logger.Debug("Current user was null.");
+            return false;
+        }
+
+        var topic = context.ApplicationMessage.Topic;
+
+        Logger.Debug("Topic was {@topic}.", topic);
+
+        if (user.ThrottleUser)
+        {
+            var payload = context.ApplicationMessage?.Payload;
+
+            if (payload is not null)
             {
-                var payload = context.ApplicationMessage?.Payload;
-
-                if (payload != null)
+                if (user.MonthlyByteLimit is not null)
                 {
-                    if (user.MonthlyByteLimit != null)
+                    if (IsUserThrottled(
+                        context.ClientId,
+                        payload.Length,
+                        user.MonthlyByteLimit.Value,
+                        dataLimitCacheMonth))
                     {
-                        if (IsUserThrottled(
-                            context.ClientId,
-                            payload.Length,
-                            user.MonthlyByteLimit.Value,
-                            dataLimitCacheMonth))
-                        {
-                            Logger.Debug("User is throttled now.");
-                            return false;
-                        }
+                        Logger.Debug("User is throttled now.");
+                        return false;
                     }
                 }
             }
+        }
 
-            Logger.Debug("The blacklist was {@blacklist}.", blacklist);
-            Logger.Debug("The whitelist was {@whitelist}.", whitelist);
+        Logger.Debug("The blacklist was {@blacklist}.", blacklist);
+        Logger.Debug("The whitelist was {@whitelist}.", whitelist);
 
-            // Check matches
-            if (blacklist.Any(b => b.Value == topic))
-            {
-                Logger.Debug("The blacklist matched a topic.");
-                return false;
-            }
-
-            if (whitelist.Any(b => b.Value == topic))
-            {
-                Logger.Debug("The whitelist matched a topic.");
-                return true;
-            }
-
-            foreach (var forbiddenTopic in blacklist)
-            {
-                var doesTopicMatch = TopicChecker.Regex(forbiddenTopic.Value, topic);
-
-                if (!doesTopicMatch)
-                {
-                    continue;
-                }
-
-                Logger.Debug("The blacklist matched a topic with regex.");
-                return false;
-            }
-
-            foreach (var allowedTopic in whitelist)
-            {
-                var doesTopicMatch = TopicChecker.Regex(allowedTopic.Value, topic);
-
-                if (!doesTopicMatch)
-                {
-                    continue;
-                }
-
-                Logger.Debug("The whitelist matched a topic with regex.");
-                return true;
-            }
-
-            Logger.Warning("We fell through everything else. This should never happen! Context was {@context}.", context);
+        // Check matches
+        if (blacklist.Any(b => b.Value == topic))
+        {
+            Logger.Debug("The blacklist matched a topic.");
             return false;
         }
 
-        /// <inheritdoc cref="IMqttValidator"/>
-        /// <summary>
-        ///     Validates the subscription.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        /// <param name="blacklist">The blacklist.</param>
-        /// <param name="whitelist">The whitelist.</param>
-        /// <param name="user">The user.</param>
-        /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
-        /// <returns>A value indicating whether the subscription is accepted or not.</returns>
-        /// <seealso cref="IMqttValidator"/>
-        public bool ValidateSubscription(
-            MqttSubscriptionInterceptorContext context,
-            List<BlacklistWhitelist> blacklist,
-            List<BlacklistWhitelist> whitelist,
-            User user,
-            List<string> clientIdPrefixes)
+        if (whitelist.Any(b => b.Value == topic))
         {
-            Logger.Debug("Executed ValidateSubscription with parameters: {@context}, {@user}.", context, user);
+            Logger.Debug("The whitelist matched a topic.");
+            return true;
+        }
 
-            var clientIdPrefix = GetClientIdPrefix(context.ClientId, clientIdPrefixes);
+        foreach (var forbiddenTopic in blacklist)
+        {
+            var doesTopicMatch = TopicChecker.Regex(forbiddenTopic.Value, topic);
 
-            Logger.Debug("Client id prefix is {@clientIdPrefix}.", clientIdPrefix);
-
-            if (user == null)
+            if (!doesTopicMatch)
             {
-                Logger.Debug("Current user was null.");
-                return false;
+                continue;
             }
 
-            var topic = context.TopicFilter.Topic;
-
-            Logger.Debug("Topic was {@topic}.", topic);
-            Logger.Debug("The blacklist was {@blacklist}.", blacklist);
-            Logger.Debug("The whitelist was {@whitelist}.", whitelist);
-
-            // Check matches
-            if (blacklist.Any(b => b.Value == topic))
-            {
-                Logger.Debug("The blacklist matched a topic.");
-                return false;
-            }
-
-            if (whitelist.Any(b => b.Value == topic))
-            {
-                Logger.Debug("The whitelist matched a topic.");
-                return true;
-            }
-
-            foreach (var forbiddenTopic in blacklist)
-            {
-                var doesTopicMatch = TopicChecker.Regex(forbiddenTopic.Value, topic);
-
-                if (!doesTopicMatch)
-                {
-                    continue;
-                }
-
-                Logger.Debug("The blacklist matched a topic with regex.");
-                return false;
-            }
-
-            foreach (var allowedTopic in whitelist)
-            {
-                var doesTopicMatch = TopicChecker.Regex(allowedTopic.Value, topic);
-
-                if (!doesTopicMatch)
-                {
-                    continue;
-                }
-
-                Logger.Debug("The whitelist matched a topic with regex.");
-                return true;
-            }
-
-            Logger.Warning(
-                "We fell through everything else. This should never happen! Context was {@context}.",
-                context);
+            Logger.Debug("The blacklist matched a topic with regex.");
             return false;
         }
 
-        /// <summary>
-        ///     Gets the client identifier prefix for a client identifier if there is one or <c>null</c> else.
-        /// </summary>
-        /// <param name="clientIdentifierParam">The client identifier.</param>
-        /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
-        /// <returns>The client identifier prefix for a client identifier if there is one or <c>null</c> else.</returns>
-        private static string GetClientIdPrefix(string clientIdentifierParam, List<string> clientIdPrefixes)
+        foreach (var allowedTopic in whitelist)
         {
-            Logger.Debug("The client id parameter was {@clientIdentifierParam}.", clientIdentifierParam);
-            Logger.Debug("The client id prefixes were {@clientIdPrefixes}.", clientIdPrefixes);
+            var doesTopicMatch = TopicChecker.Regex(allowedTopic.Value, topic);
 
-            var firstOrDefaultClientIdPrefix = clientIdPrefixes.FirstOrDefault(clientIdentifierParam.StartsWith);
-            Logger.Debug("The first or default client id prefix was {@firstOrDefaultClientIdPrefix}.", firstOrDefaultClientIdPrefix);
-            return firstOrDefaultClientIdPrefix;
+            if (!doesTopicMatch)
+            {
+                continue;
+            }
+
+            Logger.Debug("The whitelist matched a topic with regex.");
+            return true;
         }
 
-        /// <summary>
-        ///     Checks whether a user has used the maximum of its publishing limit for the month or not.
-        /// </summary>
-        /// <param name="clientId">The client identifier.</param>
-        /// <param name="sizeInBytes">The message size in bytes.</param>
-        /// <param name="monthlyByteLimit">The monthly byte limit.</param>
-        /// <param name="dataLimitCacheMonth">The data limit cache for the month.</param>
-        /// <returns>A value indicating whether the user will be throttled or not.</returns>
-        private static bool IsUserThrottled(string clientId, long sizeInBytes, long monthlyByteLimit, IMemoryCache dataLimitCacheMonth)
+        Logger.Warning("We fell through everything else. This should never happen! Context was {@context}.", context);
+        return false;
+    }
+
+    /// <inheritdoc cref="IMqttValidator"/>
+    /// <summary>
+    /// Validates the subscription.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <param name="blacklist">The blacklist.</param>
+    /// <param name="whitelist">The whitelist.</param>
+    /// <param name="user">The user.</param>
+    /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
+    /// <returns>A value indicating whether the subscription is accepted or not.</returns>
+    /// <seealso cref="IMqttValidator"/>
+    public bool ValidateSubscription(
+        MqttSubscriptionInterceptorContext context,
+        List<BlacklistWhitelist> blacklist,
+        List<BlacklistWhitelist> whitelist,
+        User user,
+        List<string> clientIdPrefixes)
+    {
+        Logger.Debug("Executed ValidateSubscription with parameters: {@context}, {@user}.", context, user);
+
+        var clientIdPrefix = GetClientIdPrefix(context.ClientId, clientIdPrefixes);
+
+        Logger.Debug("Client id prefix is {@clientIdPrefix}.", clientIdPrefix);
+
+        if (user is null)
         {
-            dataLimitCacheMonth.TryGetValue(clientId, out var foundByteLimit);
+            Logger.Debug("Current user was null.");
+            return false;
+        }
 
-            if (foundByteLimit == null)
+        var topic = context.TopicFilter.Topic;
+
+        Logger.Debug("Topic was {@topic}.", topic);
+        Logger.Debug("The blacklist was {@blacklist}.", blacklist);
+        Logger.Debug("The whitelist was {@whitelist}.", whitelist);
+
+        // Check matches
+        if (blacklist.Any(b => b.Value == topic))
+        {
+            Logger.Debug("The blacklist matched a topic.");
+            return false;
+        }
+
+        if (whitelist.Any(b => b.Value == topic))
+        {
+            Logger.Debug("The whitelist matched a topic.");
+            return true;
+        }
+
+        foreach (var forbiddenTopic in blacklist)
+        {
+            var doesTopicMatch = TopicChecker.Regex(forbiddenTopic.Value, topic);
+
+            if (!doesTopicMatch)
             {
-                dataLimitCacheMonth.Set(clientId, sizeInBytes, DateTimeOffset.Now.EndOfMonth());
+                continue;
+            }
 
-                if (sizeInBytes < monthlyByteLimit)
-                {
-                    return false;
-                }
+            Logger.Debug("The blacklist matched a topic with regex.");
+            return false;
+        }
 
+        foreach (var allowedTopic in whitelist)
+        {
+            var doesTopicMatch = TopicChecker.Regex(allowedTopic.Value, topic);
+
+            if (!doesTopicMatch)
+            {
+                continue;
+            }
+
+            Logger.Debug("The whitelist matched a topic with regex.");
+            return true;
+        }
+
+        Logger.Warning(
+            "We fell through everything else. This should never happen! Context was {@context}.",
+            context);
+        return false;
+    }
+
+    /// <summary>
+    /// Gets the client identifier prefix for a client identifier if there is one or <c>null</c> else.
+    /// </summary>
+    /// <param name="clientIdentifierParam">The client identifier.</param>
+    /// <param name="clientIdPrefixes">The client identifier prefixes.</param>
+    /// <returns>The client identifier prefix for a client identifier if there is one or <c>null</c> else.</returns>
+    private static string? GetClientIdPrefix(string clientIdentifierParam, List<string> clientIdPrefixes)
+    {
+        Logger.Debug("The client id parameter was {@clientIdentifierParam}.", clientIdentifierParam);
+        Logger.Debug("The client id prefixes were {@clientIdPrefixes}.", clientIdPrefixes);
+
+        var firstOrDefaultClientIdPrefix = clientIdPrefixes.FirstOrDefault(clientIdentifierParam.StartsWith);
+        Logger.Debug("The first or default client id prefix was {@firstOrDefaultClientIdPrefix}.", firstOrDefaultClientIdPrefix);
+        return firstOrDefaultClientIdPrefix;
+    }
+
+    /// <summary>
+    /// Checks whether a user has used the maximum of its publishing limit for the month or not.
+    /// </summary>
+    /// <param name="clientId">The client identifier.</param>
+    /// <param name="sizeInBytes">The message size in bytes.</param>
+    /// <param name="monthlyByteLimit">The monthly byte limit.</param>
+    /// <param name="dataLimitCacheMonth">The data limit cache for the month.</param>
+    /// <returns>A value indicating whether the user will be throttled or not.</returns>
+    private static bool IsUserThrottled(string clientId, long sizeInBytes, long monthlyByteLimit, IMemoryCache dataLimitCacheMonth)
+    {
+        dataLimitCacheMonth.TryGetValue(clientId, out var foundByteLimit);
+
+        if (foundByteLimit is null)
+        {
+            dataLimitCacheMonth.Set(clientId, sizeInBytes, DateTimeOffset.Now.EndOfMonth());
+
+            if (sizeInBytes < monthlyByteLimit)
+            {
+                return false;
+            }
+
+            Logger.Information("The client with client identifier {@clientId} is now locked until the end of this month because it already used its data limit.", clientId);
+            return true;
+        }
+
+        try
+        {
+            var currentValue = Convert.ToInt64(foundByteLimit);
+            currentValue = checked(currentValue + sizeInBytes);
+            dataLimitCacheMonth.Remove(clientId);
+            dataLimitCacheMonth.Set(clientId, currentValue, DateTimeOffset.Now.EndOfMonth());
+
+            if (currentValue >= monthlyByteLimit)
+            {
                 Logger.Information("The client with client identifier {@clientId} is now locked until the end of this month because it already used its data limit.", clientId);
                 return true;
             }
-
-            try
-            {
-                var currentValue = Convert.ToInt64(foundByteLimit);
-                currentValue = checked(currentValue + sizeInBytes);
-                dataLimitCacheMonth.Remove(clientId);
-                dataLimitCacheMonth.Set(clientId, currentValue, DateTimeOffset.Now.EndOfMonth());
-
-                if (currentValue >= monthlyByteLimit)
-                {
-                    Logger.Information("The client with client identifier {@clientId} is now locked until the end of this month because it already used its data limit.", clientId);
-                    return true;
-                }
-            }
-            catch (OverflowException)
-            {
-                Logger.Warning("OverflowException thrown.");
-                Logger.Information("The client with client identifier {@clientId} is now locked until the end of this month because it already used its data limit.", clientId);
-                return true;
-            }
-
-            return false;
         }
+        catch (OverflowException)
+        {
+            Logger.Warning("OverflowException thrown.");
+            Logger.Information("The client with client identifier {@clientId} is now locked until the end of this month because it already used its data limit.", clientId);
+            return true;
+        }
+
+        return false;
     }
 }
